Added stbtarget test file, to be used by remote target for stb with video div
Added the possibility to specify that the result from method invocation shall be reflected, that is to be used for remote factory functions.

Added support to specify that object properties should be serialized instead of reflected, useful when the reflected object contains a lot of constans/identifiers as properties.

TODO: Reflect outgoing callbacks passed in parameters when invoking a remote method