Implemented callback mapping so that remote invocation where callbacks are passed get reflected on local side, thus the remote end can invoke the callback back through rpc.
Added serialization for result objects so that circular references are avoided as well as excessive deep recursion on objects.
Max recursion level can be specified as an option when reflecting a remote object.

In general, the serialization/callback mapping etc is rather hacky but works, so after it has been used for real remote reflection of the target system, and tested in that usage scenario, a lot of refactoring is needed to clean up the code.

With the use of Proxy object a lot of the reflection logic can be scrapped, but the browser support is extremely vauge at the moment.